backend/requirements.txt
fastapi>=0.128.8
uvicorn[standard]>=0.38.0
beanie>=2.0.0,<2.1.0
pymongo[async]>=4.16.0
motor>=3.3.0
firebase-admin>=7.1.0
python-dotenv>=1.0.0
websockets>=13.0
httpx>=0.27.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
dateparser>=1.2.0
